experiment-setup/requirements.txt
asyncssh==2.18
Flask==3.0.3
httpx==0.27.2
otii_tcp_client==1.0.10
psycopg2_binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
pandas==pandas-2.2.3
